refactor(artwork): delete the legacy reader chain, cache warmer, and provider image methods
