+++ core/artwork/wire_providers.go
@@ -5,10 +5,8 @@ import (
 )
 
 var Set = wire.NewSet(
-	NewArtwork,
 	NewService,
 	GetImageCache,
-	NewCacheWarmer,
 	NewWorker,
 	ProvideImageStore,
 )
